Use a docker volume for node_modules
Rather than writing the node_modules to the host, which is slow, write
them to a Linux filesystem via a `docker volume`.

This was tested using a developer build of Docker Desktop:

https://desktop-stage.docker.com/mac/stable/amd64/64817/Docker.dmg

Assuming it's working correctly then running this command to show the
host filesystem I/O:
```
docker run -v /run:/run djs55/grpcfuse-control:2.3.0.3 stats calls --follow
```
should show no reads or writes to the `node_modules` directory.

It's important not to delete the `node_modules` folder on the host
(which should stay empty) otherwise the VM may get confused and write to
the host again.

Signed-off-by: David Scott <dave@recoil.org>